Day 85 - Desktop App for Adding Watermarks to Images/requirements.txt
Pillow==12.1.0
tkinterdnd2==0.4.2
PyQt5==5.15.11